+++ Changelog.md
@@ -54,7 +54,7 @@
 - Fix Variable layer height max layers and progress being reported as elephant foot fixer
 - Prevent opening the same tool window multiple times
 - Allow configuring quick layout min spacing
-
+- Remote print initialize at startup option now runs an initial scan
 
 ## v0.9.2 &mdash; August 30th, 2026